# Conflicts:
#	src/main/java/com/devcourse/web2_1_dashbunny_be/feature/order/controller/OrderController.java
#	src/main/java/com/devcourse/web2_1_dashbunny_be/feature/order/service/OrderService.java
#	src/main/java/com/devcourse/web2_1_dashbunny_be/feature/order/service/OrderServiceImpl.java

+++ src/main/java/com/devcourse/web2_1_dashbunny_be/feature/user/dto/UsersStoreListResponseDto.java
@@ -16,8 +16,8 @@ public class UsersStoreListResponseDto {
   private String storeId;
   private String storeName;
   private String storeLogo;
-  private double rating;
-  private int reviewCount;
+  private Double rating;
+  private Double reviewCount;
   private Long deliveryTip;
   private String minDeliveryTime;
   private String maxDeliveryTime;
